# PERF: Optimize read_excel nrows