--- a/pandas/io/common.py
+++ b/pandas/io/common.py
@@ -184,7 +184,7 @@ def validate_header_arg(header: object) -> None:
     if header is None:
         return
     if is_integer(header):
-        assert isinstance(header, int)
+        header = cast(int, header)
         if header < 0:
             # GH 27779
             raise ValueError(
@@ -193,7 +193,7 @@ def validate_header_arg(header: object) -> None:
             )
         return
     if is_list_like(header):
-        assert isinstance(header, Sequence)
+        header = cast(Sequence, header)
         if not all(map(is_integer, header)):
             raise ValueError("header must be integer or list of integers")
         if any(i < 0 for i in header):