- [x] closes #32727
- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added an entry in the latest `doc/source/whatsnew/vX.X.X.rst` file if fixing a bug or adding a new feature.

I timed read_excel on a file with 10 columns and 1000 rows, and report the best of 10 repeats (ms) below. This change can make a modest improvement on xls and ods files, and a significant improvement on xlsx and xlsb files. When `nrows=None` this has no measurable impact on the run time.

| ext  | nrows | time (main) | time (this branch) |
|------|-------|------------|------------|
| xls  | None  |    22.4    |    22.1    |
| xls  | 10    |    21.4    |    17.0    |
| xlsx | None  |    99.1    |    99.7    |
| xlsx | 10    |    98.0    |     8.8    |
| xlsb | None  |    81.0    |    80.2    |
| xlsb | 10    |    80.2    |     4.8    |
| ods  | None  |   571      |   569      |
| ods  | 10    |   566      |   517      |

Here are the results of ```asv run -e -E existing --bench ReadExcel``` showing similar results (the benchmark spreadsheet is different than the one above).

```
[75.00%] ··· io.excel.ReadExcel.time_read_excel                              ok
[75.00%] ··· ========== ============
               engine               
             ---------- ------------
                xlrd     36.5±0.3ms 
              openpyxl   162±0.1ms  
                odf       688±5ms   
             ========== ============

[100.00%] ··· io.excel.ReadExcelNRows.time_read_excel                         ok
[100.00%] ··· ========== ============
                engine               
              ---------- ------------
                 xlrd     24.5±0.1ms 
               openpyxl   29.0±0.1ms 
                 odf       508±3ms   
              ========== ============
```