@ahawryluk 2022-04-30
> Nice gain, especially in xlsx (I think that's pretty common). Do you know why you don't see similar gains in old/xls?

Thanks! In the case of xls, xlrd parses all of the data on first read, so there's no saving in I/O time, just a small saving inside get_sheet_data. On the other hand, I was surprised that ods showed so little improvement, especially since ods and xlsx are both zip files containing xml, as far as I understand. I'm also not sure why ods is so much slower than xlsx in the first place.

Thanks for the great recommendations on this PR. I'll take a look at them in the coming days.