@ahawryluk 2022-05-07
@rhshadrach I've made the suggested changes and they were all huge improvements except maybe the type hints. I have one failing mypy error that I'm not sure how to fix. Also, in my last commit I added a bunch of `assert isinstance` statements to work around the fact that mypy doesn't understand the type narrowing of e.g. `is_integer`. There were a few similar assert statements in _base.py, but I'm not convinced that my most recent commit was an improvement. It seems like I've lost readability, conciseness, and the original benefit of `is_integer` versus `isinstance(int)`. Maybe I should use type hint `Any` in validate_header_arg? I'm open to any suggestions you have. (No rush at all.) Have a great weekend.